backend/requirements.txt
fastapi==0.95.1
uvicorn==0.22.0
supabase==1.0.3
python-dotenv==1.0.0
pydantic==1.10.7
python-multipart==0.0.6
email-validator==2.0.0
psycopg2-binary==2.9.6 